requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
boto3==1.21.7
botocore==1.24.7
certifi==2021.10.8
charset-normalizer==2.0.11
dj-database-url==0.5.0
Django==3.2
django-countries==7.2.1
django-storages==1.12.3
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
Pillow==9.0.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
stripe==2.65.0
urllib3==1.26.8
